requirements.txt
ics==0.7.1
requests
playwright
python-dotenv
